organize the code
