+++ lib/agent/Agent.py
@@ -1,6 +1,12 @@
 ''' Class to hold information regarding the Agents' beliefs and selection algorithms '''
 import abc
 
+# https://stackoverflow.com/a/11158224/3889099
+import os,sys,inspect
+currentdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
+parentdir = os.path.dirname(currentdir)
+sys.path.insert(0,parentdir) 
+
 from constants import DEFAULT_MEMORY, DEFAULT_DISCOUNT_FACTOR
 from InformationSet import InformationSet
