+++ lib/bandit/BanditAlgorithm.py
