+++ lib/bandit/StaticGreedy.py
