+++ lib/bandit/ThompsonSampling.py
